% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{desc_output}
\alias{desc_output}
\title{Create a file to describe the output
Descriptive file are stored in the directory of the file prefixed by '.d_'}
\usage{
desc_output(path, desc = NULL, plot = FALSE)
}
\arguments{
\item{path}{character file path to describe}

\item{desc}{description to add}

\item{plot}{if TRUE use last plot title as description}
}
\description{
Create a file to describe the output
Descriptive file are stored in the directory of the file prefixed by '.d_'
}
