% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{desc_readme}
\alias{desc_readme}
\title{Create a readme file, usable by results server to show description of a directory}
\usage{
desc_readme(contents, title = NULL, path = NULL)
}
\description{
Create a readme file, usable by results server to show description of a directory
}
