% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{result_desc_filters}
\alias{result_desc_filters}
\alias{desc_filters}
\title{Create a filters description}
\usage{
result_desc_filters(auto = FALSE, filters, path = NULL)

desc_filters(auto = FALSE, filters, path = NULL)
}
\arguments{
\item{auto}{use auto filtering feature (create a filter for each context tag)}

\item{filters}{list() list of filters, each filter is a result_filter object [result_filter()]}
}
\description{
Filters are used to create filter feature on the UI.
Each filter has a name, a label, and a set of values to be used as filter
Filters allows to select a set of files using the context labels of each files
}
