% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{result_desc_output}
\alias{result_desc_output}
\alias{desc_output}
\title{Create a metadata file to describe the another file}
\usage{
result_desc_output(path, desc = NULL, plot = FALSE)

desc_output(path, desc = NULL, plot = FALSE)
}
\arguments{
\item{path}{character file path to describe}

\item{desc}{description to add}

\item{plot}{if TRUE use last plot title as description}
}
\description{
Metadata file are stored in the same directory as the file with the file name prefixed with '.d_'
Metadata file is a tiny json file that can be used by a server to get more information about the context of a file (purpose, comment, ...)
If it's a plot, it can extract the plot title as a descriptive label for the file
}
