% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{result_desc_readme}
\alias{result_desc_readme}
\alias{desc_readme}
\title{Create a readme file, usable by results server to show description of a directory}
\usage{
result_desc_readme(contents, title = NULL, path = NULL)

desc_readme(contents, title = NULL, path = NULL)
}
\arguments{
\item{contents}{readme contents}

\item{title}{title of the readme}

\item{path}{file path, by default use get_current_path() to get current path from configuration}
}
\description{
Create a readme file, usable by results server to show description of a directory
}
