% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{ResultContext}
\alias{ResultContext}
\title{Create a layered set of named values for context description off a set of files
Context is a dynamic list of named values, that can be updated by layer.
A full context is the merge of all values of all layers, with the rule, last layer overrides the first one.}
\description{
Create a layered set of named values for context description off a set of files
Context is a dynamic list of named values, that can be updated by layer.
A full context is the merge of all values of all layers, with the rule, last layer overrides the first one.

Create a layered set of named values for context description off a set of files
Context is a dynamic list of named values, that can be updated by layer.
A full context is the merge of all values of all layers, with the rule, last layer overrides the first one.
}
\details{
Context is used to manage the context of a set of files, typically during the generation of them.

For example, during data analysis, a  high number of files can be generated and some can share the same context (same output, same analysis method, same step, ...)
The context can be used to manage this shared context, an create a dedicated context for each file or group of file.
The layered feature allow to restore a previous context or to work in a recursive way (using push() and pop() to manage recursion level)

The resolved context is the set of labels in the current context with all layers merged. You can see the context as the "labels" in K8s,, a label has a name and a value.

A layer can be added (push) or removed (pop)
}
\examples{
context = ResultContext$new()
context$set(step="main")
context$push() # Create a new layer
context$set(level=1)
context$resolve()
context$pop() # Get back to first layer
context$resolve()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{contexts}}{layers as list}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ResultContext$new()}}
\item \href{#method-empty}{\code{ResultContext$empty()}}
\item \href{#method-resolve}{\code{ResultContext$resolve()}}
\item \href{#method-push}{\code{ResultContext$push()}}
\item \href{#method-pop}{\code{ResultContext$pop()}}
\item \href{#method-set}{\code{ResultContext$set()}}
\item \href{#method-clone}{\code{ResultContext$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create the instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultContext$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-empty"></a>}}
\if{latex}{\out{\hypertarget{method-empty}{}}}
\subsection{Method \code{empty()}}{
Empty the context
Create an empty context with a base layer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultContext$empty()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-resolve"></a>}}
\if{latex}{\out{\hypertarget{method-resolve}{}}}
\subsection{Method \code{resolve()}}{
Get the current contexts values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultContext$resolve()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-push"></a>}}
\if{latex}{\out{\hypertarget{method-push}{}}}
\subsection{Method \code{push()}}{
Add a new context layer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultContext$push()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pop"></a>}}
\if{latex}{\out{\hypertarget{method-pop}{}}}
\subsection{Method \code{pop()}}{
Remove last context layer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultContext$pop()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set"></a>}}
\if{latex}{\out{\hypertarget{method-set}{}}}
\subsection{Method \code{set()}}{
Define context values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultContext$set(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{list of named value to define in the current layer}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResultContext$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
