% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.R
\docType{package}
\name{swResults-package}
\alias{swResults}
\alias{swResults-package}
\title{Manipulate Output Assets Description And Context}
\description{
swResults provides a set of functions to manipulate description and context of generated files static (csv, results)
These description and context are used by an external system to build dynamic UI to present the results. An example is Results Server
}
