% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{result_filter}
\alias{result_filter}
\title{Create a filter}
\usage{
result_filter(name, label, values = NULL, ...)
}
\arguments{
\item{name}{filter name, the filter name is the name of the context label to filter on}

\item{label}{filter label}

\item{values}{list of possibles values for the label}
}
\description{
Create a filter
}
